Another fix to AddressingModeFolder (this time restoring some optimizations)
The previous changes disabled the optimization that cheri-cap-index.ll
is testing for because it would break the machine verifier. I've now
changed cheriaddressingmodefolder to iterate over the basic block backwards
so that the kill state for the GPR64 is copied to the last use of the
register instead of the next one (which is not always the last).

Also added a .mir version of cheri-cap-index.ll as well to only test the
pass rather than the whole llc optimization pipeline