When coercing between a struct <-> capability, don't dive into the struct.
The previous behaviour was causing a bug in which passing a union of the
following shape (i.e. contains a struct with at least two fields, whereby at
least one of these is a capability) by value meant that only the first field
was being passed in as opposed to the entire union/struct.

union CallData {
  struct {
    char* c;
    int i;
  } Bar;
};