When coercing between a struct <-> capability, only dive into the struct if it has exactly one field that is a capability.
Previously, if a union/struct contained a capability we would dive into this
struct and pass that field in a register regardless of whether there were other
fields in the struct. This behaviour was causing a bug in which passing a union
containing a struct with at least two fields, whereby at least one of these is
a capability, by value meant that only the first field of the struct was being
passed in as opposed to the outermost union.

We now only pass in a struct capability field directly if there are no other
fields.

This entire union will now be passed in this example:

union CallData {
  struct {
    char* __capability c;
    int i;
  } Bar;
};

Whereas the capability field c will be passed directly in a register in this
case (thus Clang "dives" into the two structs):

union Deep {
  struct {
    struct {
      char* __capability c;
    } B;
  } A;
};